trash files
instead of perma deleting them with rm, this script could be aliased
as `rm` to move said files to the trash directory just the way its done
with graphical delete.

`rmc` or other such clear aliases could be made to clear conents of this
trash dir from the terminal, if required.